requirements.txt
flask
pandas
numpy
scikit-learn
joblib
fpdf
pillow
pytesseract
flask-sqlalchemy
werkzeug